{{ cookiecutter.slug }}/backend/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile backend/requirements/base.in
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
asttokens==2.4.1
    # via stack-data
async-timeout==4.0.3
    # via redis
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
boto3==1.35.42
    # via -r backend/requirements/base.in
botocore==1.35.42
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
    # via
    #   -r backend/requirements/base.in
    #   django-celery-beat
    #   django-celery-results
    #   flower
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.5
    # via django-celery-beat
decorator==5.1.1
    # via ipython
django==4.2.16
    # via
    #   -r backend/requirements/base.in
    #   django-celery-beat
    #   django-celery-results
    #   django-cors-headers
    #   django-filter
    #   django-redis
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-celery-beat==2.7.0
    # via -r backend/requirements/base.in
django-celery-results==2.5.1
    # via -r backend/requirements/base.in
django-cors-headers==4.5.0
    # via -r backend/requirements/base.in
django-environ==0.11.2
    # via -r backend/requirements/base.in
django-filter==24.3
    # via -r backend/requirements/base.in
django-redis==5.4.0
    # via -r backend/requirements/base.in
django-storages==1.14.4
    # via -r backend/requirements/base.in
django-timezone-field==7.0
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r backend/requirements/base.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r backend/requirements/base.in
drf-spectacular==0.27.2
    # via -r backend/requirements/base.in
exceptiongroup==1.2.2
    # via ipython
executing==2.1.0
    # via stack-data
flower==2.0.1
    # via -r backend/requirements/base.in
humanize==4.11.0
    # via flower
inflection==0.5.1
    # via drf-spectacular
ipython==8.28.0
    # via -r backend/requirements/base.in
jedi==0.19.1
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
kombu==5.4.2
    # via celery
matplotlib-inline==0.1.7
    # via ipython
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.0.0
    # via -r backend/requirements/base.in
prometheus-client==0.21.0
    # via flower
prompt-toolkit==3.0.48
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.10
    # via -r backend/requirements/base.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.18.0
    # via ipython
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   python-crontab
pytz==2024.2
    # via flower
pyyaml==6.0.2
    # via drf-spectacular
redis==5.1.1
    # via django-redis
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.5.1
    # via django
stack-data==0.6.3
    # via ipython
tornado==6.4.1
    # via flower
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   asgiref
    #   ipython
tzdata==2024.2
    # via
    #   celery
    #   django-celery-beat
    #   kombu
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via botocore
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
